+++ app/src/main/java/com/example/budgetplanning/data/entities/Transaction.kt
@@ -8,7 +8,7 @@ import java.time.LocalDateTime
 @Entity(tableName = "transaction")
 data class Transaction (
     @PrimaryKey(autoGenerate = true) val id: Int? = null,
-    @ColumnInfo(name="change_amount") val changeAmount: Float,
+    @ColumnInfo(name="change_amount") val changeAmount: Double,
     @ColumnInfo(name="date_time") val dateTime: LocalDateTime = LocalDateTime.now(),
     val comment: String?
 )
